adding back test_groceriesreal.py. Moved code in usim.py into datasetinsights.io.downloader.unity_simulation
