+++ datasetinsights/datasets/synthetic.py
@@ -18,8 +18,11 @@ from datasetinsights.datasets.unity_perception import (
 from datasetinsights.datasets.unity_perception.tables import SCHEMA_VERSION
 from datasetinsights.io.bbox import BBox2D
 from datasetinsights.io.download import download_file, validate_checksum
+from datasetinsights.io.downloader.unity_simulation import (
+    Downloader,
+    download_manifest,
+)
 from datasetinsights.io.exceptions import ChecksumError
-from datasetinsights.io.usim import Downloader, download_manifest
 
 from .base import Dataset
 from .exceptions import DatasetNotFoundError
