+++ datasetinsights/stats/statistics.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import datasetinsights.constants as const
 from datasetinsights.datasets.unity_perception import MetricDefinitions, Metrics
 from datasetinsights.datasets.unity_perception.tables import SCHEMA_VERSION
-from datasetinsights.io.usim import Downloader
+from datasetinsights.io.downloader.unity_simulation import Downloader
 
 logger = logging.getLogger(__name__)
