+++ tests/unity_perception/test_download.py
@@ -13,12 +13,12 @@ from datasetinsights.io.download import (
     download_file,
     validate_checksum,
 )
-from datasetinsights.io.exceptions import ChecksumError, DownloadError
-from datasetinsights.io.usim import (
+from datasetinsights.io.downloader.unity_simulation import (
     Downloader,
     FileType,
     _filter_unsuccessful_attempts,
 )
+from datasetinsights.io.exceptions import ChecksumError, DownloadError
 
 
 @pytest.fixture
@@ -58,13 +58,13 @@ def test_download_bad_request():
 
 def test_download_rows(downloader):
     n_rows = len(downloader.manifest)
-    with patch("datasetinsights.io.usim.download_file") as mocked_dl:
+    with patch("datasetinsights.io.downloader.unity_simulation.download_file") as mocked_dl:
         matched_rows = pd.Series(np.zeros(n_rows).astype(bool))
         downloaded = downloader._download_rows(matched_rows)
         assert len(downloaded) == 0
         mocked_dl.assert_not_called()
 
-    with patch("datasetinsights.io.usim.download_file") as mocked_dl:
+    with patch("datasetinsights.io.downloader.unity_simulation.download_file") as mocked_dl:
         matched_rows = pd.Series(np.ones(n_rows).astype(bool))
         downloaded = downloader._download_rows(matched_rows)
         assert len(downloaded) == n_rows
@@ -73,7 +73,7 @@ def test_download_rows(downloader):
 
 def test_download_all(downloader):
     n_rows = len(downloader.manifest)
-    with patch("datasetinsights.io.usim.download_file") as mocked_dl:
+    with patch("datasetinsights.io.downloader.unity_simulation.download_file") as mocked_dl:
         downloader.download_references()
         downloader.download_captures()
         downloader.download_metrics()
